@@ -40,14 +40,14 @@
   },
   "scripts": {
     "build": "sl-scripts build && node ./scripts/prepare-cli",
+    "postbuild": "yarn copy.rulesets --verbose",
     "build.binary": "yarn build && yarn build.manifest && cp yarn.lock dist && cd dist && yarn && pkg . --out-path ./bin",
-    "build.docs": "sl-scripts build:typedoc",
     "build.manifest": "cd dist && oclif-dev manifest",
+    "copy.rulesets": "copyfiles -u 1 \"src/**/ruleset.json\" dist",
     "commit": "git-cz",
     "lint": "tslint 'src/**/*.ts'",
     "lint.fix": "yarn lint --fix",
     "release": "sl-scripts release",
-    "release.docs": "sl-scripts release:docs",
     "release.dryRun": "sl-scripts release --dry-run --debug",
     "test": "jest",
     "test.prod": "yarn build && yarn lint && yarn test --coverage --maxWorkers=2",
@@ -59,18 +59,19 @@
     "@oclif/command": "^1.0",
     "@oclif/config": "^1.12.11",
     "@oclif/plugin-help": "^2.0",
-    "@stoplight/json": "2.x.x",
-    "@stoplight/json-ref-resolver": "1.x.x",
-    "@stoplight/types": "4.1.x",
-    "@stoplight/yaml": "2.x.x",
+    "@stoplight/json": "^2.1.0",
+    "@stoplight/json-ref-resolver": "^1.5.0",
+    "@stoplight/types": "^7.0.1",
+    "@stoplight/yaml": "^2.3.0",
     "ajv": "6.x.x",
+    "ajv-oai": "^1.1.1",
     "chalk": "^2.4.2",
     "jsonpath": "https://github.com/stoplightio/jsonpath#f1c0e9e634da2d45671b7639fa0a99afc807da17",
     "lodash": ">=4.17.5",
     "node-fetch": "2.x.x",
     "strip-ansi": "^5.2.0",
     "text-table": "^0.2.0",
-    "typescript-json-schema": "^0.37.0"
+    "typescript-json-schema": "^0.38.0"
   },
   "devDependencies": {
     "@oclif/dev-cli": "^1",
@@ -85,15 +86,19 @@
     "@types/node": "12.0.x",
     "@types/strip-ansi": "^5.2.1",
     "@types/text-table": "^0.2.0",
+    "@types/yaml": "^1.0.2",
+    "copyfiles": "^2.1.0",
     "fs-extra": "^7.0.1",
     "jest": "24.8.x",
     "jest-when": "^2.4.0",
     "nock": "^10.0.6",
     "pkg": "4.4.0",
     "ts-jest": "24.0.x",
+    "ts-node": "^8.1.0",
     "tslint": "5.16.0",
     "tslint-config-stoplight": "1.3.0",
-    "typescript": "3.4.5"
+    "typescript": "3.5.1",
+    "yaml": "^1.5.1"
   },
   "lint-staged": {
     "*.{ts,tsx}": [
