@@ -1,106 +1,8 @@
-const merge = require('lodash/merge');
-
-import { DiagnosticSeverity } from '@stoplight/types';
-import { RuleFunction, RuleType } from '../../types';
-import { commonOasRules } from '../oas';
-import * as schema from './schemas/main.json';
+import { resolve } from 'path';
+import { readRulesFromRulesets } from '../reader';
 
 export { commonOasFunctions as oas3Functions } from '../oas';
 
-export const oas3Rules = () => {
-  return merge(commonOasRules(), {
-    // specification validation
-    'oas3-schema': {
-      summary: 'Validate structure of OpenAPIv3 specification.',
-      type: RuleType.VALIDATION,
-      severity: DiagnosticSeverity.Error,
-      then: {
-        function: RuleFunction.SCHEMA,
-        functionOptions: {
-          schema,
-        },
-      },
-      tags: ['schema'],
-    },
-
-    // generic rules
-    'api-servers': {
-      summary: 'OpenAPI `servers` must be present and non-empty array.',
-      type: RuleType.STYLE,
-      given: '$',
-      then: {
-        field: 'servers',
-        function: RuleFunction.SCHEMA,
-        functionOptions: {
-          schema: {
-            items: {
-              type: 'object',
-            },
-            minItems: 1,
-            type: 'array',
-          },
-        },
-      },
-      tags: ['api'],
-    },
-    'model-description': {
-      enabled: false,
-      summary: 'Model `description` must be present and non-empty string.',
-      type: RuleType.STYLE,
-      given: '$.components.schemas[*]',
-      then: {
-        field: 'description',
-        function: RuleFunction.TRUTHY,
-      },
-    },
-    'operation-security-defined': {
-      summary: 'Operation `security` values must match a scheme defined in the `components.securitySchemes` object.',
-      type: RuleType.VALIDATION,
-      given: '$',
-      then: {
-        function: 'oasOpSecurityDefined',
-        functionOptions: {
-          schemesPath: ['components', 'securitySchemes'],
-        },
-      },
-      tags: ['operation'],
-    },
-    'server-not-example.com': {
-      enabled: false,
-      summary: 'Server URL should not point at `example.com`.',
-      type: RuleType.STYLE,
-      given: '$.servers[*]',
-      then: {
-        field: 'url',
-        function: RuleFunction.PATTERN,
-        functionOptions: {
-          notMatch: 'example.com',
-        },
-      },
-    },
-    'server-trailing-slash': {
-      summary: 'Server URL should not have a trailing slash.',
-      type: RuleType.STYLE,
-      given: '$.servers[*]',
-      then: {
-        field: 'url',
-        function: RuleFunction.PATTERN,
-        functionOptions: {
-          notMatch: '/$',
-        },
-      },
-    },
-    'valid-example': {
-      summary: 'Examples must be valid against their defined schema.',
-      type: RuleType.VALIDATION,
-      given: '$..[?(@.example)]',
-      then: {
-        function: RuleFunction.SCHEMAPATH,
-        functionOptions: {
-          field: 'example',
-          schemaPath: '$',
-        },
-      },
-    },
-  });
+export const rules = async () => {
+  return readRulesFromRulesets(resolve(__dirname, 'ruleset.json'));
 };
