@@ -1,7 +1,18 @@
 import { DiagnosticSeverity } from '@stoplight/types';
-
-import { oas2Functions, oas2Rules } from '../rulesets/oas2';
+import * as fs from 'fs';
+import * as path from 'path';
+import { oas2Functions } from '../rulesets/oas2';
+import * as oas2Ruleset from '../rulesets/oas2/ruleset.json';
+import { oas3Functions } from '../rulesets/oas3';
+import * as oas3Ruleset from '../rulesets/oas3/ruleset.json';
 import { Spectral } from '../spectral';
+import { RuleCollection } from '../types';
+
+const invalidSchema = fs.readFileSync(
+  path.join(__dirname, './__fixtures__/petstore.invalid-schema.oas3.yaml'),
+  'utf-8',
+);
+const todosInvalid = fs.readFileSync(path.join(__dirname, './__fixtures__/todos.invalid.oas2.json'), 'utf-8');
 
 const fnName = 'fake';
 const fnName2 = 'fake2';
@@ -89,7 +100,7 @@ describe('linter', () => {
       },
     });
 
-    expect(result[0]).toEqual({
+    expect(result[0]).toMatchObject({
       code: 'rule1',
       message,
       severity: DiagnosticSeverity.Warning,
@@ -136,7 +147,7 @@ describe('linter', () => {
   });
 
   test('should include parser diagnostics', async () => {
-    spectral.addRules(oas2Rules());
+    spectral.addRules(oas2Ruleset.rules as RuleCollection);
     spectral.addFunctions(oas2Functions());
 
     const responses = `openapi: 2.0.0
@@ -185,6 +196,77 @@ responses:: !!foo
           },
           severity: 0,
         },
+      ]),
+    );
+  });
+
+  test('should merge similar ajv errors', async () => {
+    spectral.addRules(oas3Ruleset.rules as RuleCollection);
+    spectral.addFunctions(oas3Functions());
+
+    const result = await spectral.run(invalidSchema);
+
+    expect(result).toEqual(
+      expect.arrayContaining([
+        expect.objectContaining({
+          code: 'oas3-schema',
+          message: 'should NOT have additional properties: type',
+          summary: 'should NOT have additional properties: type',
+          path: ['paths', '/pets', 'get', 'responses', '200', 'headers', 'header-1'],
+        }),
+        expect.objectContaining({
+          code: 'oas3-schema',
+          message: 'should match exactly one schema in oneOf',
+          summary: 'should match exactly one schema in oneOf',
+          path: ['paths', '/pets', 'get', 'responses', '200', 'headers', 'header-1'],
+        }),
+        expect.objectContaining({
+          code: 'oas3-schema',
+          message: "should have required property '$ref'",
+          summary: "should have required property '$ref'",
+          path: ['paths', '/pets', 'get', 'responses', '200', 'headers', 'header-1'],
+        }),
+      ]),
+    );
+  });
+
+  test('should report invalid schema $refs', async () => {
+    spectral.addRules(oas3Ruleset.rules as RuleCollection);
+    spectral.addFunctions(oas3Functions());
+
+    const result = await spectral.run(todosInvalid);
+
+    expect(result).toEqual(
+      expect.arrayContaining([
+        expect.objectContaining({
+          code: 'valid-example',
+          message: '"schema" property can\'t resolve reference #/parameters/missing from id #',
+          path: ['paths', '/todos/{todoId}', 'put', 'parameters', 1, 'schema'],
+        }),
+      ]),
+    );
+  });
+
+  test('should report invalid $refs', async () => {
+    spectral.addRules(oas3Ruleset.rules as RuleCollection);
+    spectral.addFunctions(oas3Functions());
+
+    const result = await spectral.run(invalidSchema);
+
+    expect(result).toEqual(
+      expect.arrayContaining([
+        expect.objectContaining({
+          code: 'invalid-ref',
+          message: "No reader defined for scheme 'file' in ref file://models/pet.yaml",
+          path: ['paths', '/pets', 'get', 'responses', '200', 'content', 'application/json', 'schema', '$ref'],
+          severity: DiagnosticSeverity.Error,
+        }),
+        expect.objectContaining({
+          code: 'invalid-ref',
+          message: "No reader defined for scheme 'file' in ref file://../common/models/error.yaml",
+          path: ['paths', '/pets', 'get', 'responses', 'default', 'content', 'application/json', 'schema', '$ref'],
+          severity: DiagnosticSeverity.Error,
+        }),
       ]),
     );
   });
