@@ -3,17 +3,20 @@ import * as fs from 'fs';
 import { resolve } from 'path';
 type SpyInstance = jest.SpyInstance;
 
-const invalidSpecPath = resolve(__dirname, '__fixtures__/openapi-3.0-no-contact.yaml');
-const validSpecPath = resolve(__dirname, '__fixtures__/openapi-3.0-valid.yaml');
+const invalidOas3SpecPath = resolve(__dirname, '__fixtures__/openapi-3.0-no-contact.yaml');
+const validOas3SpecPath = resolve(__dirname, '__fixtures__/openapi-3.0-valid.yaml');
+const oas2PetstoreSpecPath = resolve(__dirname, '../../../__tests__/__fixtures__/petstore.oas2.json');
 const validConfigPath = resolve(__dirname, '__fixtures__/config.yml');
 const outputConfigPath = resolve(__dirname, '__fixtures__/config.output.yml');
-const validCustomSpecPath = resolve(__dirname, '__fixtures__/openapi-3.0-valid-custom.yaml');
+const validCustomOas3SpecPath = resolve(__dirname, '__fixtures__/openapi-3.0-valid-custom.yaml');
 const invalidRulesetPath = resolve(__dirname, '__fixtures__/ruleset-invalid.yaml');
 const validRulesetPath = resolve(__dirname, '__fixtures__/ruleset-valid.yaml');
 const validNestedRulesetPath = resolve(__dirname, '__fixtures__/ruleset-extends-valid.yaml');
 const invalidNestedRulesetPath = resolve(__dirname, '__fixtures__/ruleset-extends-invalid.yaml');
 const validRulesetConfigPath = resolve(__dirname, '__fixtures__/config.ruleset.yml');
 const invalidRulesetConfigPath = resolve(__dirname, '__fixtures__/config.ruleset.invalid.yml');
+const standardOas3RulesetPath = resolve(__dirname, '../../../rulesets/oas3/ruleset.json');
+const standardOas2RulesetPath = resolve(__dirname, '../../../rulesets/oas2/ruleset.json');
 
 /*
  * These tests currently do not assert stderr because it doesn't seem to be
@@ -33,14 +36,14 @@ describe('lint', () => {
     describe('and the file is expected to have no warnings', () => {
       test
         .stdout()
-        .command(['lint', validSpecPath])
+        .command(['lint', validCustomOas3SpecPath])
         .it('outputs no issues', ctx => {
           expect(ctx.stdout).toContain('No errors or warnings found!');
         });
     });
 
     describe('and the file is expected to trigger oas3 warnings', () => {
-      const args = ['lint', invalidSpecPath];
+      const args = ['lint', invalidOas3SpecPath];
 
       test
         .stdout()
@@ -57,7 +60,7 @@ describe('lint', () => {
           .command([...args, '-f', 'json'])
           .it('outputs warnings in json format', ctx => {
             expect(ctx.stdout).toContain('OpenAPI `servers` must be present and non-empty array');
-            expect(ctx.stdout).toContain('"info.contact is not truthy"');
+            expect(ctx.stdout).toContain('"Info object should contain `contact` object."');
           });
       });
 
@@ -101,14 +104,14 @@ describe('lint', () => {
     describe('extends feature', () => {
       test
         .stdout()
-        .command(['lint', validCustomSpecPath, '-r', validNestedRulesetPath])
+        .command(['lint', validCustomOas3SpecPath, '-r', validNestedRulesetPath])
         .it('should extend a valid relative ruleset', ctx => {
           expect(ctx.stdout).toContain('No errors or warnings found!');
         });
 
       test
         .stdout()
-        .command(['lint', validCustomSpecPath, '-r', invalidNestedRulesetPath])
+        .command(['lint', validCustomOas3SpecPath, '-r', invalidNestedRulesetPath])
         .exit(2)
         .it('should fail trying to extend an invalid relative ruleset', ctx => {
           expect(ctx.stdout).toContain("should have required property 'given'");
@@ -125,7 +128,7 @@ describe('lint', () => {
           });
         })
         .stdout()
-        .command(['lint', validCustomSpecPath, '-r', 'http://foo.local/ruleset-master.yaml'])
+        .command(['lint', validCustomOas3SpecPath, '-r', 'http://foo.local/ruleset-master.yaml'])
         .it('given remote nested ruleset should resolve', ctx => {
           expect(ctx.stdout).toContain('No errors or warnings found!');
         });
@@ -134,7 +137,7 @@ describe('lint', () => {
     describe('when multiple ruleset options provided', () => {
       test
         .stdout()
-        .command(['lint', validSpecPath, '-r', invalidRulesetPath, '-r', validRulesetPath])
+        .command(['lint', validOas3SpecPath, '-r', invalidRulesetPath, '-r', validRulesetPath])
         .exit(2)
         .it('given one is valid other is not, outputs "invalid ruleset" error', ctx => {
           expect(ctx.stdout).toContain(`/rules/rule-without-given-nor-them 	 should have required property 'given'`);
@@ -144,27 +147,18 @@ describe('lint', () => {
             `/rules/rule-with-invalid-enum/severity 	 should be equal to one of the allowed values`,
           );
         });
-
-      test
-        .stdout()
-        .command(['lint', validSpecPath, '-r', invalidRulesetPath, '-r', validRulesetPath])
-        .exit(2)
-        .it('given one is valid other is not, reads both', ctx => {
-          expect(ctx.stdout).toContain(`Reading ruleset ${invalidRulesetPath}`);
-          expect(ctx.stdout).toContain(`Reading ruleset ${validRulesetPath}`);
-        });
     });
 
     describe('when single ruleset option provided', () => {
       test
         .stdout()
-        .command(['lint', validSpecPath, '-r', 'non-existent-path'])
+        .command(['lint', validOas3SpecPath, '-r', 'non-existent-path'])
         .exit(2)
         .it('outputs "does not exist" error');
 
       test
         .stdout()
-        .command(['lint', validSpecPath, '-r', invalidRulesetPath])
+        .command(['lint', validOas3SpecPath, '-r', invalidRulesetPath])
         .exit(2)
         .it('outputs "invalid ruleset" error', ctx => {
           expect(ctx.stdout).toContain(`/rules/rule-without-given-nor-them 	 should have required property 'given'`);
@@ -177,14 +171,14 @@ describe('lint', () => {
 
       test
         .stdout()
-        .command(['lint', validCustomSpecPath, '-r', validRulesetPath])
+        .command(['lint', validCustomOas3SpecPath, '-r', validRulesetPath])
         .it('outputs no issues', ctx => {
           expect(ctx.stdout).toContain('No errors or warnings found!');
         });
 
       test
         .stdout()
-        .command(['lint', validSpecPath, '-r', validRulesetPath])
+        .command(['lint', validOas3SpecPath, '-r', validRulesetPath])
         .it('outputs warnings in default format', ctx => {
           expect(ctx.stdout).toContain('5:10  warning  info-matches-stoplight  Info must contain Stoplight');
           expect(ctx.stdout).not.toContain('Info object should contain `contact` object');
@@ -198,9 +192,36 @@ describe('lint', () => {
           }),
         )
         .stdout()
-        .command(['lint', validCustomSpecPath, '-r', 'http://foo.local/ruleset.yaml'])
+        .command(['lint', validCustomOas3SpecPath, '-r', 'http://foo.local/ruleset.yaml'])
         .it('given valid remote ruleset file, outputs no issues', ctx => {
           expect(ctx.stdout).toContain('No errors or warnings found!');
+        });
+    });
+
+    describe('when a standard oas3 ruleset provided through option', () => {
+      test
+        .stdout()
+        .command(['lint', invalidOas3SpecPath, '-r', standardOas3RulesetPath])
+        .it('outputs warnings in default format', ctx => {
+          expect(ctx.stdout).toContain(
+            '1:5  warning  api-servers       OpenAPI `servers` must be present and non-empty array',
+          );
+          expect(ctx.stdout).toContain('3:6  warning  info-contact      Info object should contain `contact` object');
+          expect(ctx.stdout).toContain(
+            '3:6  warning  info-description  OpenAPI object info `description` must be present and non-empty string',
+          );
+          expect(ctx.stdout).not.toContain('OpenAPI 3.x detected');
+        });
+
+      test
+        .stdout()
+        .command(['lint', oas2PetstoreSpecPath, '-r', standardOas2RulesetPath])
+        .it('outputs warnings in default format', ctx => {
+          expect(ctx.stdout).toContain(
+            '46:24  warning  operation-description   Operation `description` must be present and non-empty string',
+          );
+          expect(ctx.stdout).toContain('22 problems (0 errors, 22 warnings, 0 infos)');
+          expect(ctx.stdout).not.toContain('OpenAPI 2.x detected');
         });
     });
   });
@@ -208,7 +229,7 @@ describe('lint', () => {
   describe('when loading specification files from web', () => {
     test
       .nock('http://foo.local', api =>
-        api.get('/openapi').replyWithFile(200, validSpecPath, {
+        api.get('/openapi').replyWithFile(200, validOas3SpecPath, {
           'Content-Type': 'application/yaml',
         }),
       )
@@ -227,7 +248,7 @@ describe('lint', () => {
 
     test
       .nock('http://foo.local', api =>
-        api.get('/openapi').replyWithFile(200, invalidSpecPath, {
+        api.get('/openapi').replyWithFile(200, invalidOas3SpecPath, {
           'Content-Type': 'application/yaml',
         }),
       )
@@ -241,16 +262,16 @@ describe('lint', () => {
   describe('when using config file', () => {
     test
       .stdout()
-      .command(['lint', invalidSpecPath, '-c', validConfigPath])
+      .command(['lint', invalidOas3SpecPath, '-c', validConfigPath])
       .it('outputs warnings in json format', ctx => {
-        expect(ctx.stdout).toContain('"info.contact is not truthy"');
-        expect(ctx.stdout).toContain('"info.description is not truthy"');
-        expect(ctx.stdout).toContain('"servers does not exist"');
+        expect(ctx.stdout).toContain('"Info object should contain `contact` object."');
+        expect(ctx.stdout).toContain('"OpenAPI object info `description` must be present and non-empty string."');
+        expect(ctx.stdout).toContain('"OpenAPI `servers` must be present and non-empty array."');
       });
 
     test
       .stdout()
-      .command(['lint', invalidSpecPath, '-c', outputConfigPath])
+      .command(['lint', invalidOas3SpecPath, '-c', outputConfigPath])
       .it('saves results to a file', () => {
         expect(fs.writeFile).toHaveBeenCalledWith(
           'results.json',
@@ -262,14 +283,14 @@ describe('lint', () => {
 
     test
       .stdout()
-      .command(['lint', validSpecPath, '-c', validRulesetConfigPath])
+      .command(['lint', validOas3SpecPath, '-c', validRulesetConfigPath])
       .it('outputs invalid ruleset error when invalid ruleset provided', ctx => {
         expect(ctx.stdout).toContain(`5:10  warning  info-matches-stoplight  Info must contain Stoplight`);
       });
 
     test
       .stdout()
-      .command(['lint', validSpecPath, '-c', invalidRulesetConfigPath])
+      .command(['lint', validOas3SpecPath, '-c', invalidRulesetConfigPath])
       .exit(2)
       .it('outputs invalid ruleset error when invalid ruleset provided', ctx => {
         expect(ctx.stdout).toContain(`/rules/rule-without-given-nor-them 	 should have required property 'given'`);
@@ -279,18 +300,18 @@ describe('lint', () => {
   describe('when using config file and command args', () => {
     test
       .stdout()
-      .command(['lint', invalidSpecPath, '-c', validConfigPath, '--max-results', '1'])
+      .command(['lint', invalidOas3SpecPath, '-c', validConfigPath, '--max-results', '1'])
       .it('setting --max-results to 1 will override config value of 5', ctx => {
-        expect(ctx.stdout).toContain('"info.contact is not truthy"');
-        expect(ctx.stdout).not.toContain('"info.description is not truthy"');
-        expect(ctx.stdout).not.toContain('"servers does not exist"');
+        expect(ctx.stdout).toContain('"Info object should contain `contact` object."');
+        expect(ctx.stdout).not.toContain('"OpenAPI object info `description` must be present and non-empty string."');
+        expect(ctx.stdout).not.toContain('"OpenAPI `servers` must be present and non-empty array."');
       });
   });
 
   describe('when not using config nor default config file', () => {
     test
       .stdout()
-      .command(['lint', invalidSpecPath])
+      .command(['lint', invalidOas3SpecPath])
       .it('outputs warnings in default format', ctx => {
         expect(ctx.stdout).toContain('3:6  warning  info-contact      Info object should contain `contact` object');
       });
@@ -305,9 +326,9 @@ describe('lint', () => {
 
     test
       .stdout()
-      .command(['lint', invalidSpecPath])
+      .command(['lint', invalidOas3SpecPath])
       .it('outputs data in format from default config file', ctx => {
-        expect(ctx.stdout).toContain('"info.contact is not truthy"');
+        expect(ctx.stdout).toContain('"Info object should contain `contact` object."');
       });
   });
 });
