@@ -86,12 +86,7 @@ export const stylish = (results: IRuleResult[]): string => {
         messageType = chalk.yellow('white');
       }
 
-      return [
-        formatRange(result.range),
-        messageType,
-        result.code !== undefined ? result.code : '',
-        result.summary ? result.summary.replace(/([^ ])\.$/u, '$1') : result.message,
-      ];
+      return [formatRange(result.range), messageType, result.code !== undefined ? result.code : '', result.message];
     });
 
     output += `${table(pathTableData, {
