@@ -1,6 +1,8 @@
 import { decodePointerFragment } from '@stoplight/json';
 import * as AJV from 'ajv';
 import * as jsonSpecv4 from 'ajv/lib/refs/json-schema-draft-04.json';
+const oasFormatValidator = require('ajv-oai/lib/format-validator');
+import { IFunction, IFunctionResult, ISchemaOptions } from '../types';
 
 const ajv = new AJV({
   meta: false,
@@ -14,7 +16,31 @@ ajv._opts.defaultMeta = jsonSpecv4.id;
 // @ts-ignore
 ajv._refs['http://json-schema.org/schema'] = 'http://json-schema.org/draft-04/schema';
 
-import { IFunction, IFunctionResult, ISchemaOptions } from '../types';
+ajv.addFormat('int32', { type: 'number', validate: oasFormatValidator.int32 });
+ajv.addFormat('int64', { type: 'number', validate: oasFormatValidator.int64 });
+ajv.addFormat('float', { type: 'number', validate: oasFormatValidator.float });
+ajv.addFormat('double', { type: 'number', validate: oasFormatValidator.double });
+ajv.addFormat('byte', { type: 'string', validate: oasFormatValidator.byte });
+
+const formatPath = (path: string) =>
+  path
+    .split('/')
+    .slice(1)
+    .map(decodePointerFragment);
+
+const mergeErrors = (existingError: IFunctionResult, newError: AJV.ErrorObject) => {
+  switch (newError.keyword) {
+    case 'additionalProperties': {
+      const { additionalProperty } = newError.params as AJV.AdditionalPropertiesParams;
+      if (!new RegExp(`[:,] ${additionalProperty}`).test(existingError.message)) {
+        existingError.message += `, ${(newError.params as AJV.AdditionalPropertiesParams).additionalProperty}`;
+      }
+      return true;
+    }
+    default:
+      return existingError.message === newError.message;
+  }
+};
 
 export const schema: IFunction<ISchemaOptions> = (targetVal, opts, paths) => {
   const results: IFunctionResult[] = [];
@@ -31,25 +57,44 @@ export const schema: IFunction<ISchemaOptions> = (targetVal, opts, paths) => {
 
   const { schema: schemaObj } = opts;
 
-  // TODO: potential performance improvements (compile, etc)?
-  if (!ajv.validate(schemaObj, targetVal) && ajv.errors) {
-    ajv.errors.forEach((e: AJV.ErrorObject) => {
-      // @ts-ignore
-      if (e.params && e.params.additionalProperty) {
-        // @ts-ignore
-        e.message = e.message + ': ' + e.params.additionalProperty;
-      }
+  try {
+    if (!ajv.validate(schemaObj, targetVal) && ajv.errors) {
+      // TODO: potential performance improvements (compile, etc)?
+      const collectedErrors: string[] = [];
 
+      for (const error of ajv.errors) {
+        if (collectedErrors.length > 0) {
+          const index = collectedErrors.indexOf(error.keyword);
+          if (index !== -1) {
+            if (mergeErrors(results[index], error)) continue;
+          }
+        }
+
+        let message = error.message || '';
+
+        if (
+          error.keyword === 'additionalProperties' &&
+          (error.params as AJV.AdditionalPropertiesParams).additionalProperty
+        ) {
+          message += `: ${(error.params as AJV.AdditionalPropertiesParams).additionalProperty}`;
+        }
+
+        collectedErrors.push(error.keyword);
+        results.push({
+          path: [...path, ...formatPath(error.dataPath)],
+          message,
+        });
+      }
+    }
+  } catch (ex) {
+    if (ex instanceof AJV.MissingRefError) {
       results.push({
-        path: path.concat(
-          e.dataPath
-            .split('/')
-            .slice(1)
-            .map(frag => decodePointerFragment(frag)),
-        ),
-        message: e.message ? e.message : '',
+        message: ex.message,
+        path,
       });
-    });
+    } else {
+      throw ex;
+    }
   }
 
   return results;
