@@ -1,13 +1,13 @@
-import { Spectral } from '../../../spectral';
-import { commonOasRules } from '../index';
-
-const ruleset = { rules: commonOasRules() };
+import { DiagnosticSeverity } from '@stoplight/types';
+import { RuleType, Spectral } from '../../../spectral';
+import * as ruleset from '../ruleset.json';
 
 describe('only-local-references', () => {
   const s = new Spectral();
   s.addRules({
     'only-local-references': Object.assign(ruleset.rules['only-local-references'], {
       enabled: true,
+      type: RuleType[ruleset.rules['only-local-references'].type],
     }),
   });
 
@@ -17,7 +17,11 @@ describe('only-local-references', () => {
       paths: {},
       parameters: [{ $ref: '#/reference/path' }],
     });
-    expect(results.length).toEqual(0);
+    expect(results).toEqual([
+      expect.objectContaining({
+        code: 'invalid-ref',
+      }),
+    ]);
   });
 
   test('return errors if not local ref', async () => {
@@ -26,6 +30,14 @@ describe('only-local-references', () => {
       paths: {},
       parameters: [{ $ref: 'https://stoplight.io#/reference/path' }],
     });
-    expect(results).toMatchSnapshot();
+    expect(results).toEqual(
+      expect.arrayContaining([
+        expect.objectContaining({
+          code: 'only-local-references',
+          message: 'References should start with `#/`.',
+          severity: DiagnosticSeverity.Warning,
+        }),
+      ]),
+    );
   });
 });
