@@ -1,7 +1,5 @@
-import { Spectral } from '../../../spectral';
-import { commonOasRules } from '../index';
-
-const ruleset = { rules: commonOasRules() };
+import { RuleType, Spectral } from '../../../spectral';
+import * as ruleset from '../ruleset.json';
 
 describe('parameter-description', () => {
   const s = new Spectral();
@@ -9,6 +7,7 @@ describe('parameter-description', () => {
   s.addRules({
     'parameter-description': Object.assign(ruleset.rules['parameter-description'], {
       enabled: true,
+      type: RuleType[ruleset.rules['parameter-description'].type],
     }),
   });
 
@@ -81,10 +80,9 @@ describe('parameter-description', () => {
     expect(results).toEqual([
       expect.objectContaining({
         code: 'parameter-description',
-        message: 'parameters.limit.description is not truthy',
+        message: 'Parameter objects should have a `description`.',
         path: ['parameters', 'limit', 'description'],
         severity: 1,
-        summary: 'Parameter objects should have a `description`.',
       }),
     ]);
   });
@@ -107,10 +105,9 @@ describe('parameter-description', () => {
     expect(results).toEqual([
       expect.objectContaining({
         code: 'parameter-description',
-        message: 'paths./todos.parameters.0.description is not truthy',
+        message: 'Parameter objects should have a `description`.',
         path: ['paths', '/todos', 'parameters', 0, 'description'],
         severity: 1,
-        summary: 'Parameter objects should have a `description`.',
       }),
     ]);
   });
@@ -135,27 +132,27 @@ describe('parameter-description', () => {
     expect(results).toEqual([
       expect.objectContaining({
         code: 'parameter-description',
-        message: 'paths./todos.get.parameters.0.description is not truthy',
+        message: 'Parameter objects should have a `description`.',
         path: ['paths', '/todos', 'get', 'parameters', 0, 'description'],
         severity: 1,
-        summary: 'Parameter objects should have a `description`.',
       }),
     ]);
   });
 
-  test('does not throw on refs', async () => {
-    const results = await s.run({
-      swagger: '2.0',
-      paths: {
-        '/todos': {
-          parameters: [
-            {
-              $ref: '#/parameters/limit',
-            },
-          ],
+  test('does not throw on refs', () => {
+    return expect(
+      s.run({
+        swagger: '2.0',
+        paths: {
+          '/todos': {
+            parameters: [
+              {
+                $ref: '#/parameters/limit',
+              },
+            ],
+          },
         },
-      },
-    });
-    expect(results.length).toEqual(0);
+      }),
+    ).not.rejects;
   });
 });
