@@ -1,9 +1,11 @@
-import { Spectral } from '../../../../../index';
-import { oas2Functions, oas2Rules } from '../../../../oas2/index';
-import { oas3Functions, oas3Rules } from '../../../../oas3/index';
+import { RuleType, Spectral } from '../../../../../index';
+import { oas2Functions } from '../../../../oas2/index';
+import { rules as oas2Rules } from '../../../../oas2/ruleset.json';
+import { oas3Functions } from '../../../../oas3/index';
+import { rules as oas3Rules } from '../../../../oas3/ruleset.json';
 
-const oas2Ruleset = { functions: oas2Functions(), rules: oas2Rules() };
-const oas3Ruleset = { functions: oas3Functions(), rules: oas3Rules() };
+const oas2Ruleset = { functions: oas2Functions(), rules: oas2Rules };
+const oas3Ruleset = { functions: oas3Functions(), rules: oas3Rules };
 
 describe('oasOpSecurityDefined', () => {
   describe('oas2', () => {
@@ -12,6 +14,7 @@ describe('oasOpSecurityDefined', () => {
     s.addRules({
       'operation-security-defined': Object.assign(oas2Ruleset.rules['operation-security-defined'], {
         enabled: true,
+        type: RuleType[oas2Ruleset.rules['operation-security-defined'].type],
       }),
     });
 
@@ -61,6 +64,7 @@ describe('oasOpSecurityDefined', () => {
     s.addRules({
       'operation-security-defined': Object.assign(oas3Ruleset.rules['operation-security-defined'], {
         enabled: true,
+        type: RuleType[oas3Ruleset.rules['operation-security-defined'].type],
       }),
     });
 
