@@ -1,13 +1,12 @@
-import { Spectral } from '../../../spectral';
-import { oas2Rules } from '../index';
-
-const ruleset = { rules: oas2Rules() };
+import { RuleType, Spectral } from '../../../spectral';
+import * as ruleset from '../ruleset.json';
 
 describe('host-trailing-slash', () => {
   const s = new Spectral();
   s.addRules({
     'host-trailing-slash': Object.assign(ruleset.rules['host-trailing-slash'], {
       enabled: true,
+      type: RuleType[ruleset.rules['host-trailing-slash'].type],
     }),
   });
 
