@@ -6,8 +6,11 @@ export type Rule = IRule | TruthyRule | XorRule | LengthRule | AlphaRule | Patte
 export interface IRule<T = string, O = any> {
   type?: RuleType;
 
-  // A short summary of the rule and its intended purpose
+  // @deprecated - use message instead
   summary?: string;
+
+  // A meaningful feedback about the error
+  message?: string;
 
   // A long-form description of the rule formatted in markdown
   description?: string;
